Maintain the mapping for more than 1 URL in the localstorage. - PR requests (gitignore, move mappings.js)
